refactor(decorator): controller => view, service => controller
