--- src/main/decorator/service.ts
@@ -1,8 +1,0 @@
-import { Injectable } from './injector';
-
-/**
- * Class Decorator.
- *
- * Alias of `Injectable`, decorate a service.
- */
-export const Service = Injectable;
